fix: move AppLayout to presentation
